refactor: extract Task class and add travel time percentage support
