+++ zedd-app/src/components/AppGui.tsx
@@ -22,7 +22,8 @@ import { useEffect, useMemo } from 'react'
 import ReactMarkdown from 'react-markdown'
 
 import { ErrorBoundary } from './ErrorBoundary'
-import { AppState, Task } from '../AppState'
+import { AppState } from '../AppState'
+import { Task } from '../entities/Task'
 import { PlatformState } from '../PlatformState'
 import { ChangeSliceTaskDialog } from './ChangeSliceTaskDialog'
 import { RenameTaskDialog } from './RenameTaskDialog'
