+++ zedd-app/src/components/PlatformView.tsx
@@ -53,6 +53,7 @@ import {
   sum,
   useClasses,
   hashStringToInt,
+  getActivityPercentage
 } from '../util'
 import KeyboardArrowDownIcon from '@mui/icons-material/KeyboardArrowDown'
 import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp'
@@ -343,7 +344,7 @@ export const PlatformView = observer((props: PlatformViewProps) => {
   ).filter((taskToShow) => isTaskVisible(taskToShow))
   const theme = useTheme()
   const classes = useClasses(styles)
-  const showingTotal = sum(allWorkEntries.map((we) => we.hours))
+  const showingTotal = sum(allWorkEntries.map((we) => we.hours * getActivityPercentage(we.taskActivity)))
 
   const projectTasksViewItems: PlatformExportFormat = {}
   const ottTaskMissingRepliconTask: WorkEntry[] = []
@@ -606,7 +607,7 @@ function ProjectRow({
       .filter((we) => we.id === task.id)
 
   const getWorkedHours = (interval: Interval, task: WorkEntry) =>
-    sum(getWorkEntries(interval, task).map((we) => we.hours))
+    sum(getWorkEntries(interval, task).map((we) => we.hours * getActivityPercentage(we.taskActivity)))
 
   return (
     <>
