+++ zedd-app/src/components/TaskEditor.tsx
@@ -10,7 +10,8 @@ import { observer } from 'mobx-react-lite'
 import * as React from 'react'
 import { useCallback, useRef, useState } from 'react'
 import { InvalidPlattformUrlException, PlatformType } from 'zedd-platform'
-import { AppState, Task } from '../AppState'
+import { AppState } from '../AppState'
+import {Task} from '../entities/Task'
 import { PlatformActionType, PlatformState } from '../PlatformState'
 import { PlatformTaskSelect } from './PlatformTaskSelect'
 import { LoadingSpinner } from './LoadingSpinner'
